requirements.txt
fastapi
uvicorn
aiohttp
python-multipart
psycopg2-binary
